{SofaDefaultType] move DataTypeInfo and VirtualTypeInfo in their own .h
DataTypeInfo moves to DataTypeInfo.h
while VirtualTypeInfo is renamed to DataTypeInfoDynamicWrapper. The renaming was done to to make more clear
its role is to generate a dynamic version of a DataTypeInfo.
